@@ -1,6 +1,6 @@
 module github.com/rl404/shimakaze
 
-go 1.20
+go 1.21
 
 require (
 	github.com/PuerkitoBio/goquery v1.8.1
