@@ -2,13 +2,11 @@ package cron
 
 import (
 	"context"
-	"fmt"
-	"runtime/debug"
-	"strings"
 
-	"github.com/rl404/shimakaze/internal/errors"
+	"github.com/rl404/fairy/errors"
 	"github.com/rl404/shimakaze/internal/service"
 	"github.com/rl404/shimakaze/internal/utils"
+	"github.com/rl404/shimakaze/pkg/log"
 )
 
 // Cron contains functions for cron.
@@ -24,12 +22,11 @@ func New(service service.Service) *Cron {
 }
 
 func (c *Cron) log(ctx context.Context) {
-	if rvr := recover(); rvr != nil {
-		errors.Wrap(ctx, fmt.Errorf("%v", rvr), fmt.Errorf("%s", debug.Stack()))
-	}
-
 	errStack := errors.Get(ctx)
 	if len(errStack) > 0 {
-		utils.Error(strings.Join(errStack, ","))
+		utils.Log(map[string]interface{}{
+			"level": log.ErrorLevel,
+			"error": errStack,
+		})
 	}
 }
