@@ -5,9 +5,10 @@ import (
 	"net/http"
 
 	"github.com/rl404/fairy/cache"
+	"github.com/rl404/fairy/errors"
 	"github.com/rl404/shimakaze/internal/domain/agency/entity"
 	"github.com/rl404/shimakaze/internal/domain/agency/repository"
-	"github.com/rl404/shimakaze/internal/errors"
+	_errors "github.com/rl404/shimakaze/internal/errors"
 	"github.com/rl404/shimakaze/internal/utils"
 )
 
@@ -38,7 +39,7 @@ func (c *Cache) GetByID(ctx context.Context, id int64) (data *entity.Agency, cod
 	}
 
 	if err := c.cacher.Set(ctx, key, data); err != nil {
-		return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalCache, err)
+		return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalCache)
 	}
 
 	return data, code, nil
@@ -69,7 +70,7 @@ func (c *Cache) GetAll(ctx context.Context, req entity.GetAllRequest) (_ []entit
 	}
 
 	if err := c.cacher.Set(ctx, key, data); err != nil {
-		return nil, 0, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalCache, err)
+		return nil, 0, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalCache)
 	}
 
 	return data.Data, data.Total, code, nil
@@ -89,7 +90,7 @@ func (c *Cache) UpdateByID(ctx context.Context, id int64, data entity.Agency) (i
 
 	key := utils.GetKey("agency", id)
 	if err := c.cacher.Delete(ctx, key); err != nil {
-		return http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalCache, err)
+		return http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalCache)
 	}
 
 	return code, nil
@@ -113,7 +114,7 @@ func (c *Cache) GetCount(ctx context.Context) (data int, code int, err error) {
 	}
 
 	if err := c.cacher.Set(ctx, key, data); err != nil {
-		return 0, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalCache, err)
+		return 0, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalCache)
 	}
 
 	return data, code, nil
