@@ -4,7 +4,8 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/rl404/shimakaze/internal/errors"
+	"github.com/rl404/fairy/errors"
+	_errors "github.com/rl404/shimakaze/internal/errors"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
@@ -25,10 +26,10 @@ func New(db *mongo.Database) *Mongo {
 // Create to create non-vtuber.
 func (m *Mongo) Create(ctx context.Context, id int64) (int, error) {
 	if _, err := m.db.DeleteOne(ctx, bson.M{"id": id}); err != nil {
-		return http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalDB, err)
+		return http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalDB)
 	}
 	if _, err := m.db.InsertOne(ctx, &nonVtuber{ID: id}); err != nil {
-		return http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalDB, err)
+		return http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalDB)
 	}
 	return http.StatusCreated, nil
 }
@@ -38,14 +39,14 @@ func (m *Mongo) GetAllIDs(ctx context.Context) ([]int64, int, error) {
 	var ids []int64
 	c, err := m.db.Find(ctx, bson.M{}, options.Find().SetProjection(bson.M{"id": 1}))
 	if err != nil {
-		return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalDB, err)
+		return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalDB)
 	}
 	defer c.Close(ctx)
 
 	for c.Next(ctx) {
 		var nonVtuber nonVtuber
 		if err := c.Decode(&nonVtuber); err != nil {
-			return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalDB, err)
+			return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalDB)
 		}
 		ids = append(ids, nonVtuber.ID)
 	}
