@@ -2,11 +2,12 @@ package client
 
 import (
 	"context"
-	_errors "errors"
+	__errors "errors"
 	"net/http"
 
 	"github.com/nicklaw5/helix/v2"
-	"github.com/rl404/shimakaze/internal/errors"
+	"github.com/rl404/fairy/errors"
+	_errors "github.com/rl404/shimakaze/internal/errors"
 )
 
 // GetFollowerCount to get follower count.
@@ -21,9 +22,9 @@ func (c *Client) GetFollowerCount(ctx context.Context, id string) (int, int, err
 	})
 	if err != nil {
 		if resp == nil {
-			return 0, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+			return 0, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 		}
-		return 0, resp.StatusCode, errors.Wrap(ctx, _errors.New(resp.Error), _errors.New(resp.ErrorMessage))
+		return 0, resp.StatusCode, errors.Wrap(ctx, __errors.New(resp.Error), __errors.New(resp.ErrorMessage))
 	}
 
 	return resp.Data.Total, http.StatusOK, nil
