@@ -2,11 +2,12 @@ package client
 
 import (
 	"context"
-	_errors "errors"
+	__errors "errors"
 	"net/http"
 	"time"
 
-	"github.com/rl404/shimakaze/internal/errors"
+	"github.com/rl404/fairy/errors"
+	_errors "github.com/rl404/shimakaze/internal/errors"
 	"github.com/rl404/shimakaze/internal/utils"
 )
 
@@ -24,16 +25,16 @@ func (c *Client) setToken(ctx context.Context) (int, error) {
 	resp, err := c.client.RequestAppAccessToken([]string{})
 	if err != nil {
 		if resp == nil {
-			return http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+			return http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 		}
-		return resp.StatusCode, errors.Wrap(ctx, _errors.New(resp.Error), _errors.New(resp.ErrorMessage))
+		return resp.StatusCode, errors.Wrap(ctx, __errors.New(resp.Error), __errors.New(resp.ErrorMessage))
 	}
 
 	c.client.SetAppAccessToken(resp.Data.AccessToken)
 
 	// Save to cache.
 	if err := c.cacher.Set(ctx, key, resp.Data.AccessToken, time.Duration(resp.Data.ExpiresIn)*time.Second); err != nil {
-		return http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalCache, err)
+		return http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalCache)
 	}
 
 	return http.StatusOK, nil
