@@ -2,12 +2,13 @@ package client
 
 import (
 	"context"
-	_errors "errors"
+	__errors "errors"
 	"net/http"
 
 	"github.com/nicklaw5/helix/v2"
+	"github.com/rl404/fairy/errors"
 	"github.com/rl404/shimakaze/internal/domain/twitch/entity"
-	"github.com/rl404/shimakaze/internal/errors"
+	_errors "github.com/rl404/shimakaze/internal/errors"
 )
 
 // GetUser to get user.
@@ -19,13 +20,13 @@ func (c *Client) GetUser(ctx context.Context, name string) (*entity.User, int, e
 	resp, err := c.client.GetUsers(&helix.UsersParams{Logins: []string{name}})
 	if err != nil {
 		if resp == nil {
-			return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+			return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 		}
-		return nil, resp.StatusCode, errors.Wrap(ctx, _errors.New(resp.Error), _errors.New(resp.ErrorMessage))
+		return nil, resp.StatusCode, errors.Wrap(ctx, __errors.New(resp.Error), __errors.New(resp.ErrorMessage))
 	}
 
 	if len(resp.Data.Users) == 0 {
-		return nil, http.StatusNotFound, errors.Wrap(ctx, errors.ErrChannelNotFound)
+		return nil, http.StatusNotFound, errors.Wrap(ctx, _errors.ErrChannelNotFound)
 	}
 
 	return &entity.User{
