@@ -6,6 +6,7 @@ import (
 
 	"github.com/newrelic/go-agent/v3/newrelic"
 	"github.com/rl404/fairy/limit"
+	"github.com/rl404/fairy/limit/mutex"
 )
 
 // Client contains functions for wikia api client.
@@ -17,13 +18,12 @@ type Client struct {
 
 // New to create new wikia api client.
 func New() *Client {
-	l, _ := limit.New(limit.Mutex, 1, time.Second)
 	return &Client{
 		host: "https://virtualyoutuber.fandom.com",
 		http: &http.Client{
 			Timeout:   10 * time.Second,
 			Transport: newrelic.NewRoundTripper(http.DefaultTransport),
 		},
-		limiter: l,
+		limiter: mutex.New(1, time.Second),
 	}
 }
