@@ -3,13 +3,14 @@ package client
 import (
 	"context"
 	"encoding/json"
-	_errors "errors"
+	__errors "errors"
 	"fmt"
 	"io"
 	"net/http"
 	"net/url"
 
-	"github.com/rl404/shimakaze/internal/errors"
+	"github.com/rl404/fairy/errors"
+	_errors "github.com/rl404/shimakaze/internal/errors"
 )
 
 type getImageInfoResponse struct {
@@ -43,39 +44,39 @@ func (c *Client) GetImageInfo(ctx context.Context, name string) (string, int, er
 
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url.String(), nil)
 	if err != nil {
-		return "", http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return "", http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 	}
 
 	resp, err := c.http.Do(req)
 	if err != nil {
-		return "", http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return "", http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 	}
 	defer resp.Body.Close()
 
 	if resp.StatusCode != http.StatusOK {
-		return "", resp.StatusCode, errors.Wrap(ctx, _errors.New(http.StatusText(resp.StatusCode)))
+		return "", resp.StatusCode, errors.Wrap(ctx, __errors.New(http.StatusText(resp.StatusCode)))
 	}
 
 	respBody, err := io.ReadAll(resp.Body)
 	if err != nil {
-		return "", http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return "", http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 	}
 
 	var body getImageInfoResponse
 	if err := json.Unmarshal(respBody, &body); err != nil {
-		return "", http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return "", http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 	}
 
 	if body.Error.Info != "" {
-		return "", http.StatusBadRequest, errors.Wrap(ctx, _errors.New(body.Error.Info))
+		return "", http.StatusBadRequest, errors.Wrap(ctx, __errors.New(body.Error.Info))
 	}
 
 	if _, ok := body.Query.Pages["-1"]; ok {
-		return "", http.StatusNotFound, errors.Wrap(ctx, errors.ErrWikiaPageNotFound)
+		return "", http.StatusNotFound, errors.Wrap(ctx, _errors.ErrWikiaPageNotFound)
 	}
 
 	if len(body.Query.Pages) == 0 {
-		return "", http.StatusNotFound, errors.Wrap(ctx, errors.ErrWikiaPageNotFound)
+		return "", http.StatusNotFound, errors.Wrap(ctx, _errors.ErrWikiaPageNotFound)
 	}
 
 	for _, v := range body.Query.Pages {
@@ -86,11 +87,11 @@ func (c *Client) GetImageInfo(ctx context.Context, name string) (string, int, er
 		imgURL := v.ImageInfo[0].URL
 
 		if imgURL == "" {
-			return "", http.StatusNotFound, errors.Wrap(ctx, errors.ErrWikiaPageNotFound)
+			return "", http.StatusNotFound, errors.Wrap(ctx, _errors.ErrWikiaPageNotFound)
 		}
 
 		return imgURL, http.StatusOK, nil
 	}
 
-	return "", http.StatusNotFound, errors.Wrap(ctx, errors.ErrWikiaPageNotFound)
+	return "", http.StatusNotFound, errors.Wrap(ctx, _errors.ErrWikiaPageNotFound)
 }
