@@ -3,15 +3,16 @@ package client
 import (
 	"context"
 	"encoding/json"
-	_errors "errors"
+	__errors "errors"
 	"fmt"
 	"io"
 	"net/http"
 	"net/url"
 	"strconv"
 
+	"github.com/rl404/fairy/errors"
 	"github.com/rl404/shimakaze/internal/domain/wikia/entity"
-	"github.com/rl404/shimakaze/internal/errors"
+	_errors "github.com/rl404/shimakaze/internal/errors"
 )
 
 type getPageImageByIDResponse struct {
@@ -45,40 +46,40 @@ func (c *Client) GetPageImageByID(ctx context.Context, id int64) (*entity.PageIm
 
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url.String(), nil)
 	if err != nil {
-		return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 	}
 
 	resp, err := c.http.Do(req)
 	if err != nil {
-		return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 	}
 	defer resp.Body.Close()
 
 	if resp.StatusCode != http.StatusOK {
-		return nil, resp.StatusCode, errors.Wrap(ctx, _errors.New(http.StatusText(resp.StatusCode)))
+		return nil, resp.StatusCode, errors.Wrap(ctx, __errors.New(http.StatusText(resp.StatusCode)))
 	}
 
 	respBody, err := io.ReadAll(resp.Body)
 	if err != nil {
-		return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 	}
 
 	var body getPageImageByIDResponse
 	if err := json.Unmarshal(respBody, &body); err != nil {
-		return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 	}
 
 	if body.Error.Info != "" {
-		return nil, http.StatusBadRequest, errors.Wrap(ctx, _errors.New(body.Error.Info))
+		return nil, http.StatusBadRequest, errors.Wrap(ctx, __errors.New(body.Error.Info))
 	}
 
 	data, ok := body.Query.Pages[strconv.FormatInt(id, 10)]
 	if !ok {
-		return nil, http.StatusNotFound, errors.Wrap(ctx, errors.ErrWikiaPageNotFound)
+		return nil, http.StatusNotFound, errors.Wrap(ctx, _errors.ErrWikiaPageNotFound)
 	}
 
 	if data.Title == "" {
-		return nil, http.StatusNotFound, errors.Wrap(ctx, errors.ErrWikiaPageNotFound)
+		return nil, http.StatusNotFound, errors.Wrap(ctx, _errors.ErrWikiaPageNotFound)
 	}
 
 	return &entity.PageImage{
