@@ -3,14 +3,15 @@ package client
 import (
 	"context"
 	"encoding/json"
-	_errors "errors"
+	__errors "errors"
 	"fmt"
 	"io"
 	"net/http"
 	"net/url"
 	"time"
 
-	"github.com/rl404/shimakaze/internal/errors"
+	"github.com/rl404/fairy/errors"
+	_errors "github.com/rl404/shimakaze/internal/errors"
 )
 
 type getPlaylistItemsResponse struct {
@@ -27,7 +28,7 @@ type getPlaylistItemsResponse struct {
 func (c *Client) GetVideoIDsByChannelID(ctx context.Context, channelID string) ([]string, int, error) {
 	playlistID := c.channelIDToPlaylistID(channelID)
 	if playlistID == "" {
-		return nil, http.StatusNotFound, errors.Wrap(ctx, errors.ErrChannelNotFound)
+		return nil, http.StatusNotFound, errors.Wrap(ctx, _errors.ErrChannelNotFound)
 	}
 
 	url, _ := url.Parse(fmt.Sprintf("%s/playlistItems", c.host))
@@ -46,27 +47,27 @@ func (c *Client) GetVideoIDsByChannelID(ctx context.Context, channelID string) (
 
 		req, err := http.NewRequestWithContext(ctx, http.MethodGet, url.String(), nil)
 		if err != nil {
-			return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+			return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 		}
 
 		resp, err := c.http.Do(req)
 		if err != nil {
-			return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+			return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 		}
 		defer resp.Body.Close()
 
 		if resp.StatusCode != http.StatusOK {
-			return nil, resp.StatusCode, errors.Wrap(ctx, _errors.New(http.StatusText(resp.StatusCode)))
+			return nil, resp.StatusCode, errors.Wrap(ctx, __errors.New(http.StatusText(resp.StatusCode)))
 		}
 
 		respBody, err := io.ReadAll(resp.Body)
 		if err != nil {
-			return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+			return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 		}
 
 		var body getPlaylistItemsResponse
 		if err := json.Unmarshal(respBody, &body); err != nil {
-			return nil, http.StatusInternalServerError, errors.Wrap(ctx, errors.ErrInternalServer, err)
+			return nil, http.StatusInternalServerError, errors.Wrap(ctx, err, _errors.ErrInternalServer)
 		}
 
 		var done bool
